requirements.txt
Appium-Python-Client
alembic==1.0.11
aniso8601==4.0.1
APScheduler==3.5.3
billiard==3.5.0.4
blinker==1.4
cache==1.0.3
certifi==2018.10.15
cffi==1.11.5
chardet
Click==7.0
colorama==0.3.9
colorlog==3.1.4
concurrent-log-handler==0.9.12
configparser==3.5.0
decorator==4.3.0
et-xmlfile==1.0.1
flask_cors
Flask==1.0.2
Flask-Cache==0.13.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-APScheduler==1.11.0
PyMySQL==0.9.3
future==0.17.1
gevent
greenlet==0.4.15
har2case==0.1.11
html5lib==1.0.1
HttpRunner==2.0.6
idna==2.7
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
lxml==4.2.5
Mako==1.0.10
MarkupSafe==1.1.0
msgpack-python==0.5.6
Naked==0.1.31
openpyxl==2.5.11
passlib==1.7.1
pbr==5.1.1
pycparser==2.17
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
PyYAML==3.12
pyzmq
redis
requests
requests-toolbelt==0.8.0
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.2.14
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
tornado==5.1.1
tzlocal==1.4
urllib3==1.21.1
vine==1.1.4
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
pycryptodome==3.8.2